refactor(infra): simplify directory structure for the infra-inherited code
