+++ typescript/cli/src/rebalancer/metrics/Metrics.ts
@@ -25,15 +25,15 @@ import {
   updateNativeWalletBalanceMetrics,
   updateTokenBalanceMetrics,
   updateXERC20LimitsMetrics,
-} from './infra/scripts/warp-routes/monitor/metrics.js';
+} from './scripts/metrics.js';
 import {
   NativeWalletBalance,
   WarpRouteBalance,
   XERC20Limit,
-} from './infra/scripts/warp-routes/monitor/types.js';
-import { logger, tryFn } from './infra/scripts/warp-routes/monitor/utils.js';
-import { startMetricsServer } from './infra/src/utils/metrics.js';
-import { formatBigInt } from './utils.js';
+} from './scripts/types.js';
+import { logger, tryFn } from './scripts/utils.js';
+import { formatBigInt } from './utils/formatBigInt.js';
+import { startMetricsServer } from './utils/metrics.js';
 
 interface XERC20Info {
   limits: XERC20Limit;
