+++ typescript/cli/src/rebalancer/metrics/PriceGetter.ts
@@ -5,7 +5,7 @@ import {
   Token,
 } from '@hyperlane-xyz/sdk';
 
-import { logger } from './infra/scripts/warp-routes/monitor/utils.js';
+import { logger } from './scripts/utils.js';
 
 export class PriceGetter extends CoinGeckoTokenPriceGetter {
   private constructor({
